+++ clang/test/SemaCXX/friend-default-operator.cpp
@@ -3,10 +3,15 @@
 
 // Ensure the following out of line friend declaration doesn't cause the compiler to crash.
 
+namespace GH120857
+{
+
 class A {
   friend bool operator==(const A&, const A&);
   friend class B;
 };
 
 bool operator==(const A&, const A&) = default;
+
+}