+++ clang/docs/ReleaseNotes.rst
@@ -886,7 +886,7 @@ Bug Fixes to C++ Support
   out of a module (which is the case e.g. in MSVC's implementation of ``std`` module). (#GH118218)
 - Fixed a pack expansion issue in checking unexpanded parameter sizes. (#GH17042)
 - Fixed a bug where captured structured bindings were modifiable inside non-mutable lambda (#GH95081)
-- Fixed a crash when parsing a friend declaration and a defaulted operator.
+- Fixed a crash when parsing a friend declaration and a defaulted operator. (#GH120857)
 
 Bug Fixes to AST Handling
 ^^^^^^^^^^^^^^^^^^^^^^^^^